# perf(es/utils): Make `collect_decl` parallel